requirements.txt
pydantic
dependency-injector
pyspark
pyyaml
pyathena
pyiceberg
sqlalchemy
